[improvement][module] Change the module name to avoid the same name as the root project
