+++ Bigdata-flink-example/Flink-demo/pom.xml
@@ -8,7 +8,7 @@
         <version>1.0-SNAPSHOT</version>
     </parent>
     
-    <artifactId>Flink-cdc-example</artifactId>
+    <artifactId>Flink-demo</artifactId>
     
     <properties>
         <maven.compiler.source>8</maven.compiler.source>
