+++ Bigdata-flink-example/Flink-demo/src/main/java/io/sophiadata/flink/base/BaseInit.java
@@ -1,4 +1,4 @@
-package bigdata.flink.base;
+package io.sophiadata.flink.base;
 
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
